requirements.txt
absl-py==0.11.0
aiohttp==3.6.2
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.7
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
auto-py-to-exe==2.7.8
beautifulsoup4==4.9.1
bottle==0.12.18
bottle-websocket==0.2.9
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==3.0.7
Eel==0.12.4
filelock==3.0.12
future==0.18.2
gast==0.3.3
geographiclib==1.50
geopy==2.0.0
gevent==20.6.2
gevent-websocket==0.10.1
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
google-spreadsheet==0.0.6
googleapis-common-protos==1.52.0
googletrans==3.0.0
greenlet==0.4.16
grpcio==1.33.2
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.6.30
httpcore==0.9.1
httplib2==0.18.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.9
Keras-Preprocessing==1.1.2
Markdown==3.3.3
multidict==4.7.6
mysql-connector==2.2.9
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.0.5
pefile==2019.4.18
pipenv==2020.6.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.24.0
requests-cache==0.5.2
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
scipy==1.5.4
selenium==3.141.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
splinter==0.13.0
sqlparse==0.3.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.25
virtualenv-clone==0.5.4
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
XlsxWriter==1.2.9
yarl==1.5.1
zope.event==4.4
zope.interface==5.1.0
mysql-connector-python==8.0.12